+++ flink-runtime/src/main/java/org/apache/flink/runtime/resourcemanager/ResourceManagerRuntimeServices.java
@@ -18,8 +18,6 @@
 
 package org.apache.flink.runtime.resourcemanager;
 
-import org.apache.flink.configuration.Configuration;
-import org.apache.flink.configuration.ResourceManagerOptions;
 import org.apache.flink.runtime.concurrent.ScheduledExecutor;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.resourcemanager.slotmanager.SlotManager;
@@ -54,46 +52,6 @@ public class ResourceManagerRuntimeServices {
 	}
 
 	// -------------------- Static methods --------------------------------------
-
-	/**
-	 * Check whether the config is valid, it will throw an exception if the config
-	 * is invalid or return the cut off value.
-	 *
-	 * @param config The Flink configuration.
-	 * @param containerMemoryMB The size of the complete container, in megabytes.
-	 *
-	 * @return cut off size used by container.
-	 */
-	public static long calculateCutoffMB(Configuration config, long containerMemoryMB) {
-		Preconditions.checkArgument(containerMemoryMB > 0);
-
-		// (1) check cutoff ratio
-		final float memoryCutoffRatio = config.getFloat(
-			ResourceManagerOptions.CONTAINERIZED_HEAP_CUTOFF_RATIO);
-
-		if (memoryCutoffRatio >= 1 || memoryCutoffRatio <= 0) {
-			throw new IllegalArgumentException("The configuration value '"
-				+ ResourceManagerOptions.CONTAINERIZED_HEAP_CUTOFF_RATIO.key() + "' must be between 0 and 1. Value given="
-				+ memoryCutoffRatio);
-		}
-
-		// (2) check min cutoff value
-		final int minCutoff = config.getInteger(
-			ResourceManagerOptions.CONTAINERIZED_HEAP_CUTOFF_MIN);
-
-		if (minCutoff >= containerMemoryMB) {
-			throw new IllegalArgumentException("The configuration value '"
-				+ ResourceManagerOptions.CONTAINERIZED_HEAP_CUTOFF_MIN.key() + "'='" + minCutoff
-				+ "' is larger than the total container memory " + containerMemoryMB);
-		}
-
-		// (3) check between heap and off-heap
-		long cutoff = (long) (containerMemoryMB * memoryCutoffRatio);
-		if (cutoff < minCutoff) {
-			cutoff = minCutoff;
-		}
-		return cutoff;
-	}
 
 	public static ResourceManagerRuntimeServices fromConfiguration(
 			ResourceManagerRuntimeServicesConfiguration configuration,
