+++ flink-yarn/src/main/java/org/apache/flink/yarn/AbstractYarnClusterDescriptor.java
@@ -36,10 +36,10 @@ import org.apache.flink.configuration.SecurityOptions;
 import org.apache.flink.configuration.TaskManagerOptions;
 import org.apache.flink.runtime.akka.AkkaUtils;
 import org.apache.flink.runtime.clusterframework.BootstrapTools;
+import org.apache.flink.runtime.clusterframework.ContaineredTaskManagerParameters;
 import org.apache.flink.runtime.entrypoint.ClusterEntrypoint;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobmanager.HighAvailabilityMode;
-import org.apache.flink.runtime.resourcemanager.ResourceManagerRuntimeServices;
 import org.apache.flink.runtime.taskexecutor.TaskManagerServices;
 import org.apache.flink.util.FlinkException;
 import org.apache.flink.util.Preconditions;
@@ -411,10 +411,26 @@ public abstract class AbstractYarnClusterDescriptor implements ClusterDescriptor
 		}
 	}
 
-	private void checkConfig(ClusterSpecification clusterSpecification) {
+	/**
+	 * Method to validate cluster specification before deploy it, it will throw
+	 * an {@link IllegalConfigurationException} if the {@link ClusterSpecification} is invalid.
+	 */
+	private void validateClusterSpecification(ClusterSpecification clusterSpecification) {
 		long taskManagerMemorySize = clusterSpecification.getTaskManagerMemoryMB();
-		long cutoff = ResourceManagerRuntimeServices.calculateCutoffMB(flinkConfiguration, taskManagerMemorySize);
-		TaskManagerServices.calculateHeapSizeMB(taskManagerMemorySize - cutoff, flinkConfiguration);
+		long cutoff;
+		try {
+			// We do the validation by calling the calculation methods here
+			cutoff = ContaineredTaskManagerParameters.calculateCutoffMB(flinkConfiguration, taskManagerMemorySize);
+		} catch (IllegalArgumentException cutoffConfigurationInvalidEx) {
+			throw new IllegalConfigurationException("Configurations related to cutoff checked failed.", cutoffConfigurationInvalidEx);
+		}
+
+		try {
+			// We do the validation by calling the calculation methods here
+			TaskManagerServices.calculateHeapSizeMB(taskManagerMemorySize - cutoff, flinkConfiguration);
+		} catch (IllegalArgumentException heapSizeConfigurationInvalidEx) {
+			throw new IllegalConfigurationException("Configurations related to heap size checked failed.", heapSizeConfigurationInvalidEx);
+		}
 	}
 
 	/**
@@ -432,7 +448,7 @@ public abstract class AbstractYarnClusterDescriptor implements ClusterDescriptor
 			boolean detached) throws Exception {
 
 		// ------------------ Check if configuration is valid --------------------
-		checkConfig(clusterSpecification);
+		validateClusterSpecification(clusterSpecification);
 
 		if (UserGroupInformation.isSecurityEnabled()) {
 			// note: UGI::hasKerberosCredentials inaccurately reports false
